requirement.txt
pandas>=2.0
numpy>=1.24
tensorflow>=2.16
gensim>=4.3
scikit-learn>=1.3
tqdm>=4.66
openpyxl>=3.1
xlrd>=2.0
statsmodels>=0.14
python-docx>=1.1

